+++ src/cache/src/Driver/Driver.php
@@ -13,11 +13,14 @@ declare(strict_types=1);
 namespace Hyperf\Cache\Driver;
 
 use Hyperf\Contract\PackerInterface;
+use Hyperf\Utils\InteractsWithTime;
 use Hyperf\Utils\Packer\PhpSerializerPacker;
 use Psr\Container\ContainerInterface;
 
 abstract class Driver implements DriverInterface
 {
+    use InteractsWithTime;
+
     /**
      * @var ContainerInterface
      */