+++ src/retry/tests/RetryAnnotationAspectTest.php
@@ -12,7 +12,10 @@ declare(strict_types=1);
 
 namespace HyperfTest\Retry;
 
+use Hyperf\Contract\ContainerInterface;
 use Hyperf\Di\Aop\AnnotationMetadata;
+use Hyperf\Di\Aop\AroundInterface;
+use Hyperf\Di\Aop\Pipeline;
 use Hyperf\Di\Aop\ProceedingJoinPoint;
 use Hyperf\Di\Container;
 use Hyperf\Di\Definition\DefinitionSource;
@@ -29,6 +32,7 @@ use Hyperf\Retry\RetryBudgetInterface;
 use Hyperf\Utils\ApplicationContext;
 use Mockery;
 use PHPUnit\Framework\TestCase;
+use Swoole\Coroutine\Channel;
 use Swoole\Timer;
 
 /**
@@ -347,5 +351,62 @@ class RetryAnnotationAspectTest extends TestCase
         $point->shouldReceive('process')->andThrow(new \Exception());
         $point->shouldReceive('getArguments')->andReturns([]);
         $this->assertEquals(1, $aspect->process($point));
+    }
+
+    public function testPipeline()
+    {
+        $container = Mockery::mock(ContainerInterface::class);
+        $pipeline = new Pipeline($container);
+
+        $aspect = new RetryAnnotationAspect();
+        $aspect2 = new class() implements AroundInterface {
+            public function process(ProceedingJoinPoint $proceedingJoinPoint)
+            {
+                return $proceedingJoinPoint->process() . '_aspect';
+            }
+        };
+
+        $point = Mockery::mock(ProceedingJoinPoint::class);
+        $channel = new Channel(2);
+        $channel->push(true);
+        $channel->push(false);
+        $point->shouldReceive('processOriginalMethod')->andReturnUsing(function () use ($channel) {
+            if ($channel->pop(0.001)) {
+                throw new \Exception('broken');
+            }
+            return 'pass';
+        });
+        $point->shouldReceive('process')->andReturnUsing(function () use ($point) {
+            $closure = $point->pipe;
+            return $closure($point);
+        });
+        $point->shouldReceive('getArguments')->andReturns([]);
+        $point->shouldReceive('getAnnotationMetadata')->andReturns(
+            new class() extends AnnotationMetadata {
+                public $method;
+
+                public function __construct()
+                {
+                    $retry = new Retry();
+                    $retry->maxAttempts = 2;
+                    $retry->fallback = function () {
+                        return 'fallback';
+                    };
+                    $retry->sleepStrategyClass = FlatStrategy::class;
+                    $this->method = [
+                        AbstractRetry::class => $retry,
+                    ];
+                }
+            }
+        );
+
+        $res = $pipeline->via('process')
+            ->through([$aspect, $aspect2])
+            ->send($point)
+            ->then(function (ProceedingJoinPoint $proceedingJoinPoint) {
+                return $proceedingJoinPoint->processOriginalMethod();
+            });
+
+        $this->assertSame('pass_aspect', $res);
     }
 }