+++ src/validation/src/Concerns/ValidatesAttributes.php
@@ -161,7 +161,7 @@ trait ValidatesAttributes
             return false;
         }
 
-        return preg_match('/^[\pL\pM\pN]+$/u', $value) > 0;
+        return preg_match('/^[\pL\pM\pN]+$/u', (string) $value) > 0;
     }
 
     /**